Dashboard-Create functionality added
Now auto-creates first dashboard for user if one is not present.
Does not yet filter the loaded dashboard to a user one, that will be the next small update.
Also added ability to check session ID, and if user is not logged in, redirects them to the login page (which currently just collects an emailaddress and no pass and signs them in).