User Account Functionality Updated
Now properly filters a Dashboard for a particular user, User login and log out functionality completed, If an email is entered for a user that doesn't exist, user record is created, if user signs in and doesn't have a dashboard, dashboard is created. Special new Debuglog function added to make debugging much easier.  In future need to update the NewWidget / edit widget code to save new widgets to the specific user, this is currently broken as dashboards currently filter out just to the current user's widgets, and newwidget doesn't save the user ID.