Added Floating Bookmarks, renamed indexattempt to cachedpage.html
indexattempt.html was a placeholder for an attempt to setup a cached copy of the index.php page. This has been functioning more or less as intended for quite some time, so it's time to make it an official page. I don't want it continuing to be named something like 'index.html', as it's possible a server config error could handle the request at the root directory straight to index.html without ever loading the cache first, so i went with cachedpage.html, with is self explanatory for what it is. Other changes made are supporting ability to place a bookmark widget somewhere else on the screen aside from the left bar. If a position / size is picked, the position X/Y will be used for where to place the widget.