Turned off debug-logging
Need a clean way to enable/disable this in the future, but for now have added a flag in the code page for shared_functions to turn off debug logging. This was bloating page transfers a bit and doesn't need to be enabled all the time unless you're actively making changes.