CUSTOM WIDGET TYPE PROVIDERS
Added functionality in Setup page to support defining custom widget types and code to process them, from the setup page directly, without having to change the application files directly.
This makes use of eval(), which is dangerous and will need to have security restraints placed around it (after which point it will still be dangerous).
If this functionality is implemented, the admin may add widget types to the New Widget dialog and choose how they'll get processed.

Also - Moved several 'to-do' items to the Issues tracker in Github.

Added function getPDO_DBFile() , which returns a PDO for use with bindParam, which should enable us to properly escape characters and prevent XSS / Injections in various places that are currently at risk. Will update periodically where applicable. There is work remaining to be done on setup.php for editing / deleting a CustomWidgetProvider, but it isn't urgent so it can wait.

Bugfix applied to embed page to address globalCSS not loading.