WIP - Javascript for Widget Deletion. Not finished yet.
Added a variable 'deletebuttoncss' to widget_handling.php , and a function for trying to delete the widget to js/index.js. This code is not finished yet and doesn't execute, but should not impact the existing function.
Also resized Cancel.png to reduce unnecessary web traffic (was 512px instead of 64).