Bugfix + organizational change
Separated out code for creating new user's first dashboard into a separate php module / file, to make the load and prep file easier to read when troubleshooting.
Fixed a bug where HTMLembeds on dashboardify were being attempted to load both via javascript as well as PHP, where the javascript version seems to be broken but the php version works  , and both lay out on top of themselves without being noticed. So turned off the javascript version.
Worked on issue for default dashboard URL, but it seems to not actually update the file itself.. migt be a permissions issue.
Fixed a bug where seemingly the default dashboard URL value was getting read in with a line break after it, really annoying.